+++ CHANGELOG.md
@@ -63,6 +63,15 @@ If you want to completely restore the legacy behavior, you can disable this new 
 
 - **[FIR]**: Disallow `_` assisted context parameter names in top-level function injection.
 - **[FIR/IR]**: When generating class and provider factories now, the compiler dedupes non-assisted, non-optional injections of the same type key (i.e. type ± qualifier). This shrinks generated code size in (uncommon) scenarios where you inject the same type multiple types.
+- **[IR]**: Significantly rework the IR pipeline.
+
+Previously, Metro's IR would run in two passes:
+
+1. Collect contribution data and transform `MetroContribution` interfaces.
+2. Run all other transformations.
+
+Now, Metro runs in a single pass. Most of Metro's core transformations are run in the first full pass, collects any _seen_ dependency graphs along the way, then they are processed at the end (rather than visit the whole IR tree a second time).
+
 - **[Gradle]**: Allow `DiagnosticSeverity` metro extension properties to be configurable as `metro.*` gradle properties of the same name.
 - **[Runtime]**: Remove atomicfu dependency from Metro's core runtime artifact. The base concurrency primitives are now built on `ReentrantLock` (JVM), a port of the stdlib's Lazy spinlock on (Native), and no-op on web targets.
 
@@ -78,6 +87,7 @@ If you want to completely restore the legacy behavior, you can disable this new 
 - `enableGraphSharding` is now enabled by default. Note this only kicks in (by default) for graphs with 2000+ bindings by default.
 - Mentioned in enhancements, but worth reiterating that the underlying concurrency primitives have changed in the runtime but should be an improvement as they now use more modern reentrant locks.
 - Add Amper setup to installation docs (requires [AMPER-5095](https://youtrack.jetbrains.com/issue/AMPER-5095)).
+- Test Kotlin `2.3.20-RC`.
 
 ### Contributors